[changelog] Let bootstrapping changelog consumer consumes after-image
