+++ internal/venice-test-common/src/integrationTest/java/com/linkedin/venice/endToEnd/TestActiveActiveIngestion.java
@@ -2,7 +2,7 @@ package com.linkedin.venice.endToEnd;
 
 import static com.linkedin.davinci.store.rocksdb.RocksDBServerConfig.ROCKSDB_PLAIN_TABLE_FORMAT_ENABLED;
 import static com.linkedin.venice.CommonConfigKeys.SSL_ENABLED;
-import static com.linkedin.venice.ConfigKeys.*;
+import static com.linkedin.venice.ConfigKeys.CHILD_DATA_CENTER_KAFKA_URL_PREFIX;
 import static com.linkedin.venice.hadoop.VenicePushJobConstants.DEFAULT_KEY_FIELD_PROP;
 import static com.linkedin.venice.hadoop.VenicePushJobConstants.DEFAULT_VALUE_FIELD_PROP;
 import static com.linkedin.venice.hadoop.VenicePushJobConstants.KAFKA_INPUT_BROKER_URL;
@@ -29,40 +29,26 @@ import static com.linkedin.venice.utils.IntegrationTestPushUtils.createStoreForJ
 import static com.linkedin.venice.utils.IntegrationTestPushUtils.sendStreamingDeleteRecord;
 import static com.linkedin.venice.utils.IntegrationTestPushUtils.sendStreamingRecord;
 import static com.linkedin.venice.utils.TestUtils.generateInput;
+import static com.linkedin.venice.utils.TestWriteUtils.STRING_SCHEMA;
 import static com.linkedin.venice.utils.TestWriteUtils.getTempDataDirectory;
 
-import com.google.common.collect.ImmutableSet;
-import com.linkedin.avroutil1.compatibility.AvroCompatibilityHelper;
-import com.linkedin.davinci.consumer.BootstrappingVeniceChangelogConsumer;
-import com.linkedin.davinci.consumer.ChangeEvent;
-import com.linkedin.davinci.consumer.ChangelogClientConfig;
-import com.linkedin.davinci.consumer.VeniceAfterImageConsumerImpl;
-import com.linkedin.davinci.consumer.VeniceChangeCoordinate;
-import com.linkedin.davinci.consumer.VeniceChangelogConsumer;
-import com.linkedin.davinci.consumer.VeniceChangelogConsumerClientFactory;
 import com.linkedin.venice.ConfigKeys;
 import com.linkedin.venice.client.store.AvroGenericStoreClient;
 import com.linkedin.venice.client.store.ClientConfig;
 import com.linkedin.venice.client.store.ClientFactory;
 import com.linkedin.venice.controllerapi.ControllerClient;
-import com.linkedin.venice.controllerapi.MultiStoreTopicsResponse;
 import com.linkedin.venice.controllerapi.UpdateStoreQueryParams;
 import com.linkedin.venice.controllerapi.VersionCreationResponse;
-import com.linkedin.venice.integration.utils.PubSubBrokerConfigs;
 import com.linkedin.venice.integration.utils.PubSubBrokerWrapper;
 import com.linkedin.venice.integration.utils.ServiceFactory;
 import com.linkedin.venice.integration.utils.VeniceClusterWrapper;
 import com.linkedin.venice.integration.utils.VeniceControllerWrapper;
 import com.linkedin.venice.integration.utils.VeniceMultiClusterWrapper;
-import com.linkedin.venice.integration.utils.VeniceRouterWrapper;
 import com.linkedin.venice.integration.utils.VeniceServerWrapper;
 import com.linkedin.venice.integration.utils.VeniceTwoLayerMultiRegionMultiClusterWrapper;
-import com.linkedin.venice.integration.utils.ZkServerWrapper;
 import com.linkedin.venice.meta.VeniceUserStoreType;
 import com.linkedin.venice.meta.Version;
-import com.linkedin.venice.meta.ViewConfig;
 import com.linkedin.venice.pubsub.PubSubProducerAdapterFactory;
-import com.linkedin.venice.pubsub.api.PubSubMessage;
 import com.linkedin.venice.pushmonitor.ExecutionStatus;
 import com.linkedin.venice.samza.VeniceSystemFactory;
 import com.linkedin.venice.samza.VeniceSystemProducer;
@@ -70,13 +56,11 @@ import com.linkedin.venice.serializer.AvroSerializer;
 import com.linkedin.venice.utils.DataProviderUtils;
 import com.linkedin.venice.utils.IntegrationTestPushUtils;
 import com.linkedin.venice.utils.MockCircularTime;
-import com.linkedin.venice.utils.TestMockTime;
 import com.linkedin.venice.utils.TestUtils;
 import com.linkedin.venice.utils.TestWriteUtils;
 import com.linkedin.venice.utils.Time;
 import com.linkedin.venice.utils.Utils;
 import com.linkedin.venice.view.TestView;
-import com.linkedin.venice.views.ChangeCaptureView;
 import com.linkedin.venice.writer.VeniceWriter;
 import com.linkedin.venice.writer.VeniceWriterFactory;
 import com.linkedin.venice.writer.VeniceWriterOptions;
@@ -84,8 +68,6 @@ import io.tehuti.metrics.MetricsRepository;
 import java.io.File;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -123,8 +105,7 @@ public class TestActiveActiveIngestion {
 
   @BeforeClass(alwaysRun = true)
   public void setUp() {
-    String stringSchemaStr = "\"string\"";
-    serializer = new AvroSerializer(AvroCompatibilityHelper.parse(stringSchemaStr));
+    serializer = new AvroSerializer(STRING_SCHEMA);
     Properties serverProperties = new Properties();
     serverProperties.setProperty(ConfigKeys.SERVER_PROMOTION_TO_LEADER_REPLICA_DELAY_SECONDS, Long.toString(1));
     serverProperties.put(ROCKSDB_PLAIN_TABLE_FORMAT_ENABLED, false);
@@ -164,78 +145,6 @@ public class TestActiveActiveIngestion {
   public void cleanUp() {
     multiRegionMultiClusterWrapper.close();
     TestView.resetCounters();
-  }
-
-  private void pollChangeEventsFromChangeCaptureConsumerToList(
-      List<PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> polledChangeEvents,
-      VeniceChangelogConsumer veniceChangelogConsumer) {
-    Collection<PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> pubSubMessages =
-        veniceChangelogConsumer.poll(1000);
-    for (PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate> pubSubMessage: pubSubMessages) {
-      polledChangeEvents.add(pubSubMessage);
-    }
-  }
-
-  private void pollChangeEventsFromChangeCaptureConsumer(
-      Map<String, PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> polledChangeEvents,
-      VeniceChangelogConsumer veniceChangelogConsumer) {
-    Collection<PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> pubSubMessages =
-        veniceChangelogConsumer.poll(1000);
-    for (PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate> pubSubMessage: pubSubMessages) {
-      String key = pubSubMessage.getKey().toString();
-      polledChangeEvents.put(key, pubSubMessage);
-    }
-  }
-
-  private void pollChangeEventsFromChangeCaptureConsumer(
-      Map<String, PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> polledChangeEvents,
-      BootstrappingVeniceChangelogConsumer bootstrappingVeniceChangelogConsumer) {
-    Collection<PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> pubSubMessages =
-        bootstrappingVeniceChangelogConsumer.poll(1000);
-    for (PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate> pubSubMessage: pubSubMessages) {
-      String key = pubSubMessage.getKey() == null ? null : pubSubMessage.getKey().toString();
-      polledChangeEvents.put(key, pubSubMessage);
-    }
-  }
-
-  private void pollChangeEventsFromChangeCaptureConsumer2(
-      Map<String, PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> polledChangeEvents,
-      VeniceChangelogConsumer veniceChangelogConsumer) {
-    Collection<PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> pubSubMessages =
-        veniceChangelogConsumer.poll(1000);
-    pubSubMessages.addAll(veniceChangelogConsumer.poll(1000));
-    pubSubMessages.addAll(veniceChangelogConsumer.poll(1000));
-    for (PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate> pubSubMessage: pubSubMessages) {
-      String key = pubSubMessage.getKey().toString();
-      polledChangeEvents.put(key, pubSubMessage);
-    }
-  }
-
-  private void pollChangeEventsFromChangeCaptureConsumer2(
-      Map<String, PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> polledChangeEvents,
-      BootstrappingVeniceChangelogConsumer bootstrappingVeniceChangelogConsumer) {
-    Collection<PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> pubSubMessages =
-        bootstrappingVeniceChangelogConsumer.poll(1000);
-    pubSubMessages.addAll(bootstrappingVeniceChangelogConsumer.poll(1000));
-    pubSubMessages.addAll(bootstrappingVeniceChangelogConsumer.poll(1000));
-    for (PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate> pubSubMessage: pubSubMessages) {
-      String key = pubSubMessage.getKey().toString();
-      polledChangeEvents.put(key, pubSubMessage);
-    }
-  }
-
-  private int pollAfterImageEventsFromChangeCaptureConsumer(
-      Map<String, Utf8> polledChangeEvents,
-      VeniceChangelogConsumer veniceChangelogConsumer) {
-    int polledMessagesNum = 0;
-    Collection<PubSubMessage<Utf8, ChangeEvent<Utf8>, Long>> pubSubMessages = veniceChangelogConsumer.poll(1000);
-    for (PubSubMessage<Utf8, ChangeEvent<Utf8>, Long> pubSubMessage: pubSubMessages) {
-      Utf8 afterImageEvent = pubSubMessage.getValue().getCurrentValue();
-      String key = pubSubMessage.getKey().toString();
-      polledChangeEvents.put(key, afterImageEvent);
-      polledMessagesNum++;
-    }
-    return polledMessagesNum;
   }
 
   @Test(timeOut = TEST_TIMEOUT, dataProviderClass = DataProviderUtils.class)
@@ -569,992 +478,16 @@ public class TestActiveActiveIngestion {
     });
   }
 
-  @Test(timeOut = TEST_TIMEOUT, priority = 3)
-  public void testAAIngestionWithStoreView() throws Exception {
-    // Set up the store
-    Long timestamp = System.currentTimeMillis();
-    ControllerClient childControllerClient =
-        new ControllerClient(clusterName, childDatacenters.get(0).getControllerConnectString());
-    // create a active-active enabled store and run batch push job
-    // batch job contains 100 records
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = TestWriteUtils.writeSimpleAvroFileWithStringToStringSchema(inputDir);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String storeName = Utils.getUniqueString("store");
-    Properties props =
-        TestWriteUtils.defaultVPJProps(parentControllers.get(0).getControllerUrl(), inputDirPath, storeName);
-    String keySchemaStr = recordSchema.getField(DEFAULT_KEY_FIELD_PROP).schema().toString();
-    String valueSchemaStr = recordSchema.getField(DEFAULT_VALUE_FIELD_PROP).schema().toString();
-    Map<String, String> viewConfig = new HashMap<>();
-    props.put(KAFKA_LINGER_MS, 0);
-    viewConfig.put(
-        "testViewWrong",
-        "{\"viewClassName\" : \"" + TestView.class.getCanonicalName() + "\", \"viewParameters\" : {}}");
-    UpdateStoreQueryParams storeParms = new UpdateStoreQueryParams().setActiveActiveReplicationEnabled(true)
-        .setHybridRewindSeconds(500)
-        .setHybridOffsetLagThreshold(8)
-        .setChunkingEnabled(true)
-        .setNativeReplicationEnabled(true)
-        .setPartitionCount(3);
-    MetricsRepository metricsRepository = new MetricsRepository();
-    ControllerClient setupControllerClient =
-        createStoreForJob(clusterName, keySchemaStr, valueSchemaStr, props, storeParms);
-    UpdateStoreQueryParams storeParams1 = new UpdateStoreQueryParams().setStoreViews(viewConfig);
-    setupControllerClient
-        .retryableRequest(5, controllerClient1 -> setupControllerClient.updateStore(storeName, storeParams1));
-    UpdateStoreQueryParams storeParams2 =
-        new UpdateStoreQueryParams().setViewName("testViewWrong").setDisableStoreView();
-    setupControllerClient
-        .retryableRequest(5, controllerClient1 -> setupControllerClient.updateStore(storeName, storeParams2));
-    TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-      Map<String, ViewConfig> viewConfigMap = setupControllerClient.getStore(storeName).getStore().getViewConfigs();
-      Assert.assertTrue(viewConfigMap.isEmpty());
-    });
-
-    UpdateStoreQueryParams storeParams3 = new UpdateStoreQueryParams().setViewName("changeCaptureView")
-        .setViewClassName(ChangeCaptureView.class.getCanonicalName())
-        .setViewClassParams(Collections.singletonMap("kafka.linger.ms", "0"));
-    setupControllerClient
-        .retryableRequest(5, controllerClient1 -> setupControllerClient.updateStore(storeName, storeParams3));
-
-    UpdateStoreQueryParams storeParams4 =
-        new UpdateStoreQueryParams().setViewName("testView").setViewClassName(TestView.class.getCanonicalName());
-    setupControllerClient
-        .retryableRequest(5, controllerClient1 -> setupControllerClient.updateStore(storeName, storeParams4));
-
-    TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-      Map<String, ViewConfig> viewConfigMap = setupControllerClient.getStore(storeName).getStore().getViewConfigs();
-      Assert.assertEquals(viewConfigMap.size(), 2);
-      Assert.assertEquals(viewConfigMap.get("testView").getViewClassName(), TestView.class.getCanonicalName());
-      Assert.assertEquals(
-          viewConfigMap.get("changeCaptureView").getViewClassName(),
-          ChangeCaptureView.class.getCanonicalName());
-      Assert.assertEquals(viewConfigMap.get("changeCaptureView").getViewParameters().size(), 1);
-    });
-
-    // Write Records to the store for version v1, the push job will contain 100 records.
-    TestWriteUtils.runPushJob("Run push job", props);
-
-    // Write Records from nearline
-    Map<String, String> samzaConfig = getSamzaConfig(storeName);
-    VeniceSystemFactory factory = new VeniceSystemFactory();
-    // Use a unique key for DELETE with RMD validation
-    int deleteWithRmdKeyIndex = 1000;
-
-    TestMockTime testMockTime = new TestMockTime();
-    ZkServerWrapper localZkServer = multiRegionMultiClusterWrapper.getChildRegions().get(0).getZkServerWrapper();
-    try (PubSubBrokerWrapper localKafka = ServiceFactory.getPubSubBroker(
-        new PubSubBrokerConfigs.Builder().setZkWrapper(localZkServer)
-            .setMockTime(testMockTime)
-            .setRegionName("local-pubsub")
-            .build())) {
-      Properties consumerProperties = new Properties();
-      String localKafkaUrl = localKafka.getAddress();
-      consumerProperties.put(KAFKA_BOOTSTRAP_SERVERS, localKafkaUrl);
-      ChangelogClientConfig globalChangelogClientConfig = new ChangelogClientConfig().setViewName("changeCaptureView")
-          .setConsumerProperties(consumerProperties)
-          .setControllerD2ServiceName(D2_SERVICE_NAME)
-          .setD2ServiceName(VeniceRouterWrapper.CLUSTER_DISCOVERY_D2_SERVICE_NAME)
-          .setLocalD2ZkHosts(localZkServer.getAddress())
-          .setControllerRequestRetryCount(3)
-          .setVersionSwapDetectionIntervalTimeInMs(10L);
-      VeniceChangelogConsumerClientFactory veniceChangelogConsumerClientFactory =
-          new VeniceChangelogConsumerClientFactory(globalChangelogClientConfig, metricsRepository);
-
-      ChangelogClientConfig globalAfterImageClientConfig =
-          ChangelogClientConfig.cloneConfig(globalChangelogClientConfig).setViewName("");
-      VeniceChangelogConsumerClientFactory veniceAfterImageConsumerClientFactory =
-          new VeniceChangelogConsumerClientFactory(globalAfterImageClientConfig, metricsRepository);
-
-      VeniceChangelogConsumer<Utf8, Utf8> versionTopicConsumer =
-          veniceAfterImageConsumerClientFactory.getChangelogConsumer(storeName);
-      Assert.assertTrue(versionTopicConsumer instanceof VeniceAfterImageConsumerImpl);
-      versionTopicConsumer.subscribeAll().get();
-
-      // Let's consume those 100 records off of version 1
-      Map<String, Utf8> versionTopicEvents = new HashMap<>();
-      pollAfterImageEventsFromChangeCaptureConsumer(versionTopicEvents, versionTopicConsumer);
-      Assert.assertEquals(versionTopicEvents.size(), 100);
-
-      VeniceChangelogConsumer<Utf8, Utf8> veniceChangelogConsumer =
-          veniceChangelogConsumerClientFactory.getChangelogConsumer(storeName);
-      veniceChangelogConsumer.subscribeAll().get();
-      try (VeniceSystemProducer veniceProducer =
-          factory.getClosableProducer("venice", new MapConfig(samzaConfig), null)) {
-        veniceProducer.start();
-        // Run Samza job to send PUT and DELETE requests.
-        runSamzaStreamJob(veniceProducer, storeName, null, 10, 10, 100);
-        // Produce a DELETE record with large timestamp
-        produceRecordWithLogicalTimestamp(veniceProducer, storeName, deleteWithRmdKeyIndex, 1000, true);
-      }
-
-      try (AvroGenericStoreClient<String, Utf8> client = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName)
-              .setVeniceURL(clusterWrapper.getRandomRouterURL())
-              .setMetricsRepository(metricsRepository))) {
-        TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-          Assert.assertNull(client.get(Integer.toString(deleteWithRmdKeyIndex)).get());
-        });
-      }
-
-      // Validate change events for version 1. 100 records exist in version 1.
-      Map<String, PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> polledChangeEvents = new HashMap<>();
-      Map<String, PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> allChangeEvents = new HashMap<>();
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        Assert.assertEquals(polledChangeEvents.size(), 100);
-      });
-
-      allChangeEvents.putAll(polledChangeEvents);
-      polledChangeEvents.clear();
-
-      // 21 changes in nearline. 10 puts, 10 deletes, and 1 record with a producer timestamp
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        // 21 events for nearline events
-        Assert.assertEquals(polledChangeEvents.size(), 21);
-        for (int i = 100; i < 110; i++) {
-          String key = Integer.toString(i);
-          ChangeEvent<Utf8> changeEvent = polledChangeEvents.get(key).getValue();
-          Assert.assertNotNull(changeEvent);
-          if (i != 100) {
-            Assert.assertNull(changeEvent.getPreviousValue());
-          } else {
-            Assert.assertTrue(changeEvent.getPreviousValue().toString().contains(key));
-          }
-          Assert.assertEquals(changeEvent.getCurrentValue().toString(), "stream_" + i);
-        }
-        for (int i = 110; i < 120; i++) {
-          String key = Integer.toString(i);
-          ChangeEvent<Utf8> changeEvent = polledChangeEvents.get(key).getValue();
-          Assert.assertNotNull(changeEvent);
-          Assert.assertNull(changeEvent.getPreviousValue()); // schema id is negative, so we did not parse.
-          Assert.assertNull(changeEvent.getCurrentValue());
-        }
-      });
-
-      versionTopicEvents.clear();
-      TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-        pollAfterImageEventsFromChangeCaptureConsumer(versionTopicEvents, versionTopicConsumer);
-        Assert.assertEquals(versionTopicEvents.size(), 21);
-      });
-
-      // run repush. Repush will reapply all existing events to the new store and trim all events from the RT
-      props.setProperty(SOURCE_KAFKA, "true");
-      props.setProperty(KAFKA_INPUT_BROKER_URL, clusterWrapper.getPubSubBrokerWrapper().getAddress());
-      props.setProperty(KAFKA_INPUT_MAX_RECORDS_PER_MAPPER, "5");
-      // intentionally stop re-consuming from RT so stale records don't affect the testing results
-      props.put(REWIND_TIME_IN_SECONDS_OVERRIDE, 0);
-      TestWriteUtils.runPushJob("Run repush job", props);
-      ControllerClient controllerClient =
-          new ControllerClient(clusterName, childDatacenters.get(0).getControllerConnectString());
-      TestUtils.waitForNonDeterministicAssertion(
-          5,
-          TimeUnit.SECONDS,
-          () -> Assert.assertEquals(controllerClient.getStore(storeName).getStore().getCurrentVersion(), 2));
-      clusterWrapper.refreshAllRouterMetaData();
-      // Validate repush from version 2
-      try (AvroGenericStoreClient<String, Utf8> client = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName)
-              .setVeniceURL(clusterWrapper.getRandomRouterURL())
-              .setMetricsRepository(metricsRepository))) {
-        TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-          // test single get
-          for (int i = 100; i < 110; i++) {
-            String key = Integer.toString(i);
-            Utf8 value = client.get(key).get();
-            Assert.assertNotNull(value);
-            Assert.assertEquals(value.toString(), "stream_" + i);
-          }
-          // test deletes
-          for (int i = 110; i < 120; i++) {
-            String key = Integer.toString(i);
-            Utf8 value = client.get(key).get();
-            Assert.assertNull(value);
-          }
-          // test old data
-          for (int i = 20; i < 100; i++) {
-            String key = Integer.toString(i);
-            Utf8 value = client.get(key).get();
-            Assert.assertNotNull(value);
-            Assert.assertTrue(value.toString().contains(String.valueOf(i).substring(0, 0)));
-          }
-        });
-      }
-      try (VeniceSystemProducer veniceProducer =
-          factory.getClosableProducer("venice", new MapConfig(samzaConfig), null)) {
-        veniceProducer.start();
-        // Produce a new PUT with smaller logical timestamp, it is expected to be ignored as there was a DELETE with
-        // larger
-        // timestamp
-        produceRecordWithLogicalTimestamp(veniceProducer, storeName, deleteWithRmdKeyIndex, 2, false);
-        // Produce another record to the same partition to make sure the above PUT is processed during validation stage.
-        produceRecordWithLogicalTimestamp(veniceProducer, storeName, deleteWithRmdKeyIndex + 1, 1, false);
-      }
-      try (AvroGenericStoreClient<String, Utf8> client = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName)
-              .setVeniceURL(clusterWrapper.getRandomRouterURL())
-              .setMetricsRepository(metricsRepository))) {
-        TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-          Assert.assertNotNull(client.get(Integer.toString(deleteWithRmdKeyIndex + 1)).get());
-        });
-        TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-          Assert.assertNull(client.get(Integer.toString(deleteWithRmdKeyIndex)).get());
-        });
-      }
-
-      // Validate changed events for version 2.
-      allChangeEvents.putAll(polledChangeEvents);
-      polledChangeEvents.clear();
-      // As records keys from VPJ start from 1, real-time produced records' key starts from 0, the message with key as 0
-      // is new message.
-      TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-        // poll enough to get through the empty push and the topic jump to RT.
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        String deleteWithRmdKey = Integer.toString(deleteWithRmdKeyIndex);
-        String persistWithRmdKey = Integer.toString(deleteWithRmdKeyIndex + 1);
-        Assert.assertNull(polledChangeEvents.get(deleteWithRmdKey));
-        Assert.assertNotNull(polledChangeEvents.get(persistWithRmdKey));
-        Assert.assertEquals(
-            polledChangeEvents.get(persistWithRmdKey).getValue().getCurrentValue().toString(),
-            "stream_" + persistWithRmdKey);
-      });
-
-      /**
-       * Test Repush with TTL
-       */
-      // run empty push to clean up batch data
-      parentControllerClient.sendEmptyPushAndWait(storeName, "Run empty push job", 1000, 30 * Time.MS_PER_SECOND);
-      // set up mocked time for Samza records so some records can be stale intentionally.
-      List<Long> mockTimestampInMs = new LinkedList<>();
-      Instant now = Instant.now();
-      // always-valid record
-      mockTimestampInMs.add(now.toEpochMilli());
-      // always-stale records since ttl time is 360 sec
-      Instant past = now.minus(1, ChronoUnit.HOURS);
-      mockTimestampInMs.add(past.toEpochMilli());
-      Time mockTime = new MockCircularTime(mockTimestampInMs);
-      try (VeniceSystemProducer veniceProducer =
-          factory.getClosableProducer("venice", new MapConfig(samzaConfig), null)) {
-        veniceProducer.start();
-        // run samza to stream put and delete
-        runSamzaStreamJob(veniceProducer, storeName, mockTime, 10, 10, 20);
-      }
-      // Validate changed events for version 3.
-      // AtomicInteger totalPolledAfterImageMessages = new AtomicInteger();
-      // Map<String, Utf8> polledAfterImageEvents = new HashMap<>();
-      // Map<String, Utf8> totalPolledAfterImageEvents = new HashMap<>();
-
-      TestUtils.waitForNonDeterministicAssertion(60, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        // Filter previous 21 messages.
-        Assert.assertEquals(polledChangeEvents.size(), 1);
-      });
-
-      // Consume from beginning of the version that was current at time of consumer subscription (version 2) since
-      // version 2
-      // was a repush of 101 records (0-100) with streaming updates on 100-110 and deletes on 110-119, then we expect
-      // a grand total of 119 records in this version. We'll consume up to EOP
-
-      // TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-      // totalPolledAfterImageMessages
-      // .addAndGet(pollAfterImageEventsFromChangeCaptureConsumer(polledAfterImageEvents, veniceAfterImageConsumer));
-      // Assert.assertEquals(polledAfterImageEvents.size(), 119);
-      // totalPolledAfterImageEvents.putAll(polledAfterImageEvents);
-      // polledAfterImageEvents.clear();
-      // });
-
-      // We'll have consumed everything on version
-      // TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-      // totalPolledAfterImageMessages
-      // .addAndGet(pollAfterImageEventsFromChangeCaptureConsumer(polledAfterImageEvents, veniceAfterImageConsumer));
-      // Assert.assertEquals(polledAfterImageEvents.size(), 0);
-      // totalPolledAfterImageEvents.putAll(polledAfterImageEvents);
-      // polledAfterImageEvents.clear();
-      // });
-
-      // After image consumer consumed 3 different topics: v2, v2_cc and v3_cc.
-      // The total messages: 102 (v2 repush from v1, key: 0-100, 1000) + 1 (v2_cc, key: 1001) + 42 (v3_cc, key: 0-39,
-      // 1000, 1001) - 22 (filtered from v3_cc, key: 0-19, 1000 and 1001 as they were read already.)
-      // Assert.assertEquals(totalPolledAfterImageMessages.get(), 149);
-
-      // for (int i = 1; i < 100; i++) {
-      // String key = Integer.toString(i);
-      // Utf8 afterImageValue = totalPolledAfterImageEvents.get(key);
-      // if (i < 20) {
-      // Assert.assertNotNull(afterImageValue);
-      // Assert.assertEquals(afterImageValue.toString(), "test_name_" + i);
-      // } else if (i < 40 && i >= 30) {
-      // Deleted
-      // Assert.assertNull(afterImageValue);
-      // } else {
-      // Assert.assertTrue(afterImageValue.toString().contains(String.valueOf(i).substring(0, 0)));
-      // }
-      // }
-
-      // Drain the remaining events on version 3 and verify that we got everything. We don't verify the count
-      // because at this stage, the total events which will get polled will be determined by how far back the rewind
-      // managed to get (and test run duration might be variable)
-      TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        for (int i = 20; i < 40; i++) {
-          String key = Integer.toString(i);
-          ChangeEvent<Utf8> changeEvent = polledChangeEvents.get(key).getValue();
-          Assert.assertNotNull(changeEvent);
-          Assert.assertNull(changeEvent.getPreviousValue());
-          if (i >= 20 && i < 30) {
-            Assert.assertEquals(changeEvent.getCurrentValue().toString(), "stream_" + i);
-          } else {
-            Assert.assertNull(changeEvent.getCurrentValue());
-          }
-        }
-      });
-
-      allChangeEvents.putAll(polledChangeEvents);
-      polledChangeEvents.clear();
-
-      // This should get everything submitted to the CC topic on this version since the timestamp is before anything got
-      // transmitted
-      veniceChangelogConsumer.seekToTimestamp(timestamp);
-
-      // test pause and resume
-      veniceChangelogConsumer.pause();
-      TestUtils.waitForNonDeterministicAssertion(5, TimeUnit.SECONDS, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        Assert.assertEquals(polledChangeEvents.size(), 0);
-      });
-      veniceChangelogConsumer.resume();
-
-      // This should get everything submitted to the CC topic on this version since the timestamp is before anything got
-      // transmitted
-      TestUtils.waitForNonDeterministicAssertion(5, TimeUnit.SECONDS, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        Assert.assertEquals(polledChangeEvents.size(), 42);
-      });
-      allChangeEvents.putAll(polledChangeEvents);
-      polledChangeEvents.clear();
-
-      // enable repush ttl
-      props.setProperty(REPUSH_TTL_ENABLE, "true");
-      TestWriteUtils.runPushJob("Run repush job with TTL", props);
-      TestUtils.waitForNonDeterministicAssertion(
-          5,
-          TimeUnit.SECONDS,
-          () -> Assert.assertEquals(controllerClient.getStore(storeName).getStore().getCurrentVersion(), 4));
-      // Validate repush from version 4
-      clusterWrapper.refreshAllRouterMetaData();
-      try (AvroGenericStoreClient<String, Utf8> client = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName)
-              .setVeniceURL(clusterWrapper.getRandomRouterURL())
-              .setMetricsRepository(metricsRepository))) {
-        // test single get
-        int validGet = 0, filteredGet = 0;
-        for (int i = 20; i < 30; i++) {
-          Object result = client.get(Integer.toString(i)).get();
-          if (result == null) {
-            filteredGet++;
-          } else {
-            validGet++;
-          }
-        }
-        // Half records are valid, another half is not
-        Assert.assertEquals(validGet, 5);
-        Assert.assertEquals(filteredGet, 5);
-        // test deletes
-        for (int i = 30; i < 40; i++) {
-          // not matter the DELETE is TTLed or not, the value should always be null
-          Assert.assertNull(client.get(Integer.toString(i)).get());
-        }
-        // test old data - should be empty due to empty push
-        for (int i = 40; i < 100; i++) {
-          Assert.assertNull(client.get(Integer.toString(i)).get());
-        }
-      }
-
-      // Since nothing is produced, so no changed events generated.
-      TestUtils.waitForNonDeterministicAssertion(5, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumer2(polledChangeEvents, veniceChangelogConsumer);
-        Assert.assertEquals(polledChangeEvents.size(), 0);
-      });
-
-      // Seek to the beginning of the push
-      veniceChangelogConsumer.seekToBeginningOfPush().join();
-      TestUtils.waitForNonDeterministicAssertion(5, TimeUnit.SECONDS, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        Assert.assertEquals(polledChangeEvents.size(), 30);
-      });
-
-      // Save a checkpoint and clear the map
-      Set<VeniceChangeCoordinate> checkpointSet = new HashSet<>();
-      checkpointSet.add(polledChangeEvents.get(Integer.toString(20)).getOffset());
-      allChangeEvents.putAll(polledChangeEvents);
-      polledChangeEvents.clear();
-
-      // Seek the consumer by checkpoint
-      veniceChangelogConsumer.seekToCheckpoint(checkpointSet).join();
-      allChangeEvents.putAll(polledChangeEvents);
-      polledChangeEvents.clear();
-
-      // Poll Change events again, verify we get everything
-      TestUtils.waitForNonDeterministicAssertion(5, TimeUnit.SECONDS, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        // Repush with TTL will include delete events in the topic
-        Assert.assertEquals(polledChangeEvents.size(), 16);
-      });
-      allChangeEvents.putAll(polledChangeEvents);
-      polledChangeEvents.clear();
-      Assert.assertEquals(allChangeEvents.size(), 121);
-
-      // Seek the consumer to the beginning of push (since the latest is version 4 with no nearline writes, shouldn't
-      // have any new writes)
-      // veniceAfterImageConsumer.seekToEndOfPush().join();
-      TestUtils.waitForNonDeterministicAssertion(5, TimeUnit.SECONDS, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        Assert.assertEquals(polledChangeEvents.size(), 0);
-      });
-
-      // Also should be nothing on the tail
-      // veniceAfterImageConsumer.seekToTail().join();
-      TestUtils.waitForNonDeterministicAssertion(5, TimeUnit.SECONDS, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        Assert.assertEquals(polledChangeEvents.size(), 0);
-      });
-
-      // This should get everything submitted to the CC topic on this version (version 4 doesn't have anything)
-      veniceChangelogConsumer.seekToTimestamp(timestamp);
-      TestUtils.waitForNonDeterministicAssertion(5, TimeUnit.SECONDS, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, veniceChangelogConsumer);
-        Assert.assertEquals(polledChangeEvents.size(), 0);
-      });
-
-      versionTopicEvents.clear();
-      TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-        pollAfterImageEventsFromChangeCaptureConsumer(versionTopicEvents, versionTopicConsumer);
-        // At this point, the consumer should have auto tracked to version 4, and since we didn't apply any nearline
-        // writes to version 4, there should be no events to consume at this point
-        Assert.assertEquals(versionTopicEvents.size(), 0);
-      });
-
-      versionTopicConsumer.seekToEndOfPush().get();
-      TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-        pollAfterImageEventsFromChangeCaptureConsumer(versionTopicEvents, versionTopicConsumer);
-        // Again, no events to consume here.
-        Assert.assertEquals(versionTopicEvents.size(), 0);
-      });
-
-      versionTopicConsumer.seekToBeginningOfPush().get();
-      TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-        pollAfterImageEventsFromChangeCaptureConsumer(versionTopicEvents, versionTopicConsumer);
-        // Reconsuming the events from the version topic, which at this point should just contain the same 16
-        // events we consumed with the before/after image consumer ealier.
-        Assert.assertEquals(versionTopicEvents.size(), 16);
-      });
-
-      // Verify version swap count matches with version count - 1 (since we don't transmit from version 0 to version 1).
-      // This will include messages for all partitions, so (4 version -1)*3 partitions=9 messages
-      TestUtils.waitForNonDeterministicAssertion(
-          5,
-          TimeUnit.SECONDS,
-          () -> Assert.assertEquals(TestView.getInstance().getVersionSwapCountForStore(storeName), 9));
-      // Verify total updates match up (first 20 + next 20 should make 40, And then double it again as rewind updates
-      // are
-      // applied to a version)
-      TestUtils.waitForNonDeterministicAssertion(
-          8,
-          TimeUnit.SECONDS,
-          () -> Assert.assertEquals(TestView.getInstance().getRecordCountForStore(storeName), 86));
-      parentControllerClient.disableAndDeleteStore(storeName);
-      // Verify that topics and store is cleaned up
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-        MultiStoreTopicsResponse storeTopicsResponse = childControllerClient.getDeletableStoreTopics();
-        Assert.assertFalse(storeTopicsResponse.isError());
-        Assert.assertEquals(storeTopicsResponse.getTopics().size(), 0);
-      });
-    }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT, priority = 3)
-  public void testBootstrappingVeniceChangelogConsumer_WithoutPollAndCatchUp() throws Exception {
-    ControllerClient childControllerClient =
-        new ControllerClient(clusterName, childDatacenters.get(0).getControllerConnectString());
-    // create a active-active enabled store and run batch push job
-    // batch job contains 100 records
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = TestWriteUtils.writeSimpleAvroFileWithStringToStringSchema(inputDir);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String storeName = Utils.getUniqueString("store");
-    Properties props =
-        TestWriteUtils.defaultVPJProps(parentControllers.get(0).getControllerUrl(), inputDirPath, storeName);
-    String keySchemaStr = recordSchema.getField(DEFAULT_KEY_FIELD_PROP).schema().toString();
-    String valueSchemaStr = recordSchema.getField(DEFAULT_VALUE_FIELD_PROP).schema().toString();
-    Map<String, String> viewConfig = new HashMap<>();
-    viewConfig.put(
-        "testView",
-        "{\"viewClassName\" : \"" + TestView.class.getCanonicalName() + "\", \"viewParameters\" : {}}");
-
-    viewConfig.put(
-        "changeCaptureView",
-        "{\"viewClassName\" : \"" + ChangeCaptureView.class.getCanonicalName() + "\", \"viewParameters\" : {}}");
-    UpdateStoreQueryParams storeParms = new UpdateStoreQueryParams().setActiveActiveReplicationEnabled(true)
-        .setHybridRewindSeconds(500)
-        .setHybridOffsetLagThreshold(8)
-        .setChunkingEnabled(true)
-        .setNativeReplicationEnabled(true)
-        .setPartitionCount(3);
-    MetricsRepository metricsRepository = new MetricsRepository();
-    ControllerClient setupControllerClient =
-        createStoreForJob(clusterName, keySchemaStr, valueSchemaStr, props, storeParms);
-    storeParms.setStoreViews(viewConfig);
-    // IntegrationTestPushUtils.updateStore(clusterName, props, storeParms);
-    setupControllerClient
-        .retryableRequest(5, controllerClient1 -> setupControllerClient.updateStore(storeName, storeParms));
-    // controllerClient.updateStore(storeName, storeParms);
-    TestWriteUtils.runPushJob("Run push job", props);
-    Map<String, String> samzaConfig = getSamzaConfig(storeName);
-    VeniceSystemFactory factory = new VeniceSystemFactory();
-    // Use a unique key for DELETE with RMD validation
-    int deleteWithRmdKeyIndex = 1000;
-
-    TestMockTime testMockTime = new TestMockTime();
-    ZkServerWrapper localZkServer = multiRegionMultiClusterWrapper.getChildRegions().get(0).getZkServerWrapper();
-    try (PubSubBrokerWrapper localKafka = ServiceFactory.getPubSubBroker(
-        new PubSubBrokerConfigs.Builder().setZkWrapper(localZkServer)
-            .setMockTime(testMockTime)
-            .setRegionName("local-pubsub")
-            .build())) {
-      Properties consumerProperties = new Properties();
-      String localKafkaUrl = localKafka.getAddress();
-      consumerProperties.put(KAFKA_BOOTSTRAP_SERVERS, localKafkaUrl);
-      consumerProperties.put(CLUSTER_NAME, clusterName);
-      consumerProperties.put(ZOOKEEPER_ADDRESS, localZkServer.getAddress());
-      ChangelogClientConfig globalChangelogClientConfig = new ChangelogClientConfig().setViewName("changeCaptureView")
-          .setConsumerProperties(consumerProperties)
-          .setControllerD2ServiceName(D2_SERVICE_NAME)
-          .setD2ServiceName(VeniceRouterWrapper.CLUSTER_DISCOVERY_D2_SERVICE_NAME)
-          .setLocalD2ZkHosts(localZkServer.getAddress())
-          .setControllerRequestRetryCount(3)
-          .setBootstrapFileSystemPath(Utils.getUniqueString(inputDirPath));
-      VeniceChangelogConsumerClientFactory veniceChangelogConsumerClientFactory =
-          new VeniceChangelogConsumerClientFactory(globalChangelogClientConfig, metricsRepository);
-
-      BootstrappingVeniceChangelogConsumer<Utf8, Utf8> bootstrappingVeniceChangelogConsumer =
-          veniceChangelogConsumerClientFactory.getBootstrappingChangelogConsumer(storeName);
-      bootstrappingVeniceChangelogConsumer.start().get();
-      try (VeniceSystemProducer veniceProducer =
-          factory.getClosableProducer("venice", new MapConfig(samzaConfig), null)) {
-        veniceProducer.start();
-        // Run Samza job to send PUT and DELETE requests.
-        runSamzaStreamJob(veniceProducer, storeName, null, 10, 10, 100);
-        // Produce a DELETE record with large timestamp
-        produceRecordWithLogicalTimestamp(veniceProducer, storeName, deleteWithRmdKeyIndex, 1000, true);
-      }
-
-      try (AvroGenericStoreClient<String, Utf8> client = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName)
-              .setVeniceURL(clusterWrapper.getRandomRouterURL())
-              .setMetricsRepository(metricsRepository))) {
-        TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-          Assert.assertNull(client.get(Integer.toString(deleteWithRmdKeyIndex)).get());
-        });
-      }
-
-      Map<String, PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> polledChangeEvents = new HashMap<>();
-
-      // 21 changes in nearline. 10 puts, 10 deletes, and 1 record with a producer timestamp
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, bootstrappingVeniceChangelogConsumer);
-        // 21 events for nearline events
-        Assert.assertEquals(polledChangeEvents.size(), 22);
-        for (int i = 100; i < 110; i++) {
-          String key = Integer.toString(i);
-          ChangeEvent<Utf8> changeEvent = polledChangeEvents.get(key).getValue();
-          Assert.assertNotNull(changeEvent);
-          if (i != 100) {
-            Assert.assertNull(changeEvent.getPreviousValue());
-          } else {
-            Assert.assertTrue(changeEvent.getPreviousValue().toString().contains(key));
-          }
-          Assert.assertEquals(changeEvent.getCurrentValue().toString(), "stream_" + i);
-        }
-        for (int i = 110; i < 120; i++) {
-          String key = Integer.toString(i);
-          ChangeEvent<Utf8> changeEvent = polledChangeEvents.get(key).getValue();
-          Assert.assertNotNull(changeEvent);
-          Assert.assertNull(changeEvent.getPreviousValue()); // schema id is negative, so we did not parse.
-          Assert.assertNull(changeEvent.getCurrentValue());
-        }
-      });
-
-      polledChangeEvents.clear();
-
-      // Since nothing is produced, so no changed events generated.
-      TestUtils.waitForNonDeterministicAssertion(5, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumer2(polledChangeEvents, bootstrappingVeniceChangelogConsumer);
-        Assert.assertEquals(polledChangeEvents.size(), 0);
-      });
-
-      // Verify total updates match up
-      TestUtils.waitForNonDeterministicAssertion(
-          5,
-          TimeUnit.SECONDS,
-          () -> Assert.assertEquals(TestView.getInstance().getRecordCountForStore(storeName), 21));
-      parentControllerClient.disableAndDeleteStore(storeName);
-      // Verify that topics and store is cleaned up
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-        MultiStoreTopicsResponse storeTopicsResponse = childControllerClient.getDeletableStoreTopics();
-        Assert.assertFalse(storeTopicsResponse.isError());
-        Assert.assertEquals(storeTopicsResponse.getTopics().size(), 0);
-      });
-    }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT, priority = 3)
-  public void testBootstrappingVeniceChangelogConsumer_WithPollAndCatchUp() throws Exception {
-    ControllerClient childControllerClient =
-        new ControllerClient(clusterName, childDatacenters.get(0).getControllerConnectString());
-    // create a active-active enabled store and run batch push job
-    // batch job contains 100 records
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = TestWriteUtils.writeSimpleAvroFileWithStringToStringSchema(inputDir);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String storeName = Utils.getUniqueString("store");
-    Properties props =
-        TestWriteUtils.defaultVPJProps(parentControllers.get(0).getControllerUrl(), inputDirPath, storeName);
-    String keySchemaStr = recordSchema.getField(DEFAULT_KEY_FIELD_PROP).schema().toString();
-    String valueSchemaStr = recordSchema.getField(DEFAULT_VALUE_FIELD_PROP).schema().toString();
-    Map<String, String> viewConfig = new HashMap<>();
-    viewConfig.put(
-        "testView",
-        "{\"viewClassName\" : \"" + TestView.class.getCanonicalName() + "\", \"viewParameters\" : {}}");
-
-    viewConfig.put(
-        "changeCaptureView",
-        "{\"viewClassName\" : \"" + ChangeCaptureView.class.getCanonicalName() + "\", \"viewParameters\" : {}}");
-    UpdateStoreQueryParams storeParms = new UpdateStoreQueryParams().setActiveActiveReplicationEnabled(true)
-        .setHybridRewindSeconds(500)
-        .setHybridOffsetLagThreshold(8)
-        .setChunkingEnabled(true)
-        .setNativeReplicationEnabled(true)
-        .setPartitionCount(3);
-    MetricsRepository metricsRepository = new MetricsRepository();
-    ControllerClient setupControllerClient =
-        createStoreForJob(clusterName, keySchemaStr, valueSchemaStr, props, storeParms);
-    storeParms.setStoreViews(viewConfig);
-    // IntegrationTestPushUtils.updateStore(clusterName, props, storeParms);
-    setupControllerClient
-        .retryableRequest(5, controllerClient1 -> setupControllerClient.updateStore(storeName, storeParms));
-    // controllerClient.updateStore(storeName, storeParms);
-    TestWriteUtils.runPushJob("Run push job", props);
-    Map<String, String> samzaConfig = getSamzaConfig(storeName);
-    VeniceSystemFactory factory = new VeniceSystemFactory();
-    // Use a unique key for DELETE with RMD validation
-    int deleteWithRmdKeyIndex = 1000;
-
-    TestMockTime testMockTime = new TestMockTime();
-    ZkServerWrapper localZkServer = multiRegionMultiClusterWrapper.getChildRegions().get(0).getZkServerWrapper();
-    try (PubSubBrokerWrapper localKafka = ServiceFactory.getPubSubBroker(
-        new PubSubBrokerConfigs.Builder().setZkWrapper(localZkServer)
-            .setMockTime(testMockTime)
-            .setRegionName("local-pubsub")
-            .build())) {
-      Properties consumerProperties = new Properties();
-      String localKafkaUrl = localKafka.getAddress();
-      consumerProperties.put(KAFKA_BOOTSTRAP_SERVERS, localKafkaUrl);
-      consumerProperties.put(CLUSTER_NAME, clusterName);
-      consumerProperties.put(ZOOKEEPER_ADDRESS, localZkServer.getAddress());
-      ChangelogClientConfig globalChangelogClientConfig = new ChangelogClientConfig().setViewName("changeCaptureView")
-          .setConsumerProperties(consumerProperties)
-          .setControllerD2ServiceName(D2_SERVICE_NAME)
-          .setD2ServiceName(VeniceRouterWrapper.CLUSTER_DISCOVERY_D2_SERVICE_NAME)
-          .setLocalD2ZkHosts(localZkServer.getAddress())
-          .setControllerRequestRetryCount(3)
-          .setBootstrapFileSystemPath(Utils.getUniqueString(inputDirPath));
-      VeniceChangelogConsumerClientFactory veniceChangelogConsumerClientFactory =
-          new VeniceChangelogConsumerClientFactory(globalChangelogClientConfig, metricsRepository);
-
-      BootstrappingVeniceChangelogConsumer<Utf8, Utf8> bootstrappingVeniceChangelogConsumer =
-          veniceChangelogConsumerClientFactory.getBootstrappingChangelogConsumer(storeName);
-      try (VeniceSystemProducer veniceProducer =
-          factory.getClosableProducer("venice", new MapConfig(samzaConfig), null)) {
-        veniceProducer.start();
-        // Run Samza job to send PUT and DELETE requests.
-        runSamzaStreamJob(veniceProducer, storeName, null, 10, 10, 100);
-        // Produce a DELETE record with large timestamp
-        produceRecordWithLogicalTimestamp(veniceProducer, storeName, deleteWithRmdKeyIndex, 1000, true);
-      }
-
-      try (AvroGenericStoreClient<String, Utf8> client = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName)
-              .setVeniceURL(clusterWrapper.getRandomRouterURL())
-              .setMetricsRepository(metricsRepository))) {
-        TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-          Assert.assertNull(client.get(Integer.toString(deleteWithRmdKeyIndex)).get());
-        });
-      }
-
-      // Wait for 10 seconds so bootstrap can load results from kafka
-      Utils.sleep(10000);
-      bootstrappingVeniceChangelogConsumer.start().get();
-      Map<String, PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> polledChangeEvents = new HashMap<>();
-
-      // 21 changes in nearline. 10 puts, 10 deletes and 1 delete with timestamp will be ignored and 1 end of bootstrap
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, bootstrappingVeniceChangelogConsumer);
-        Assert.assertEquals(polledChangeEvents.size(), 11);
-
-        for (int i = 100; i < 110; i++) {
-          String key = Integer.toString(i);
-          ChangeEvent<Utf8> changeEvent = polledChangeEvents.get(key).getValue();
-          Assert.assertNotNull(changeEvent);
-          Assert.assertNull(changeEvent.getPreviousValue());
-          Assert.assertEquals(changeEvent.getCurrentValue().toString(), "stream_" + i);
-        }
-
-        Assert.assertTrue(polledChangeEvents.containsKey(null));
-        Assert.assertTrue(polledChangeEvents.get(null).isEndOfBootstrap());
-      });
-
-      polledChangeEvents.clear();
-
-      // Since nothing is produced, so no changed events generated.
-      TestUtils.waitForNonDeterministicAssertion(5, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumer2(polledChangeEvents, bootstrappingVeniceChangelogConsumer);
-        Assert.assertEquals(polledChangeEvents.size(), 0);
-      });
-
-      // Verify total updates match up
-      TestUtils.waitForNonDeterministicAssertion(
-          5,
-          TimeUnit.SECONDS,
-          () -> Assert.assertEquals(TestView.getInstance().getRecordCountForStore(storeName), 21));
-      parentControllerClient.disableAndDeleteStore(storeName);
-      // Verify that topics and store is cleaned up
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-        MultiStoreTopicsResponse storeTopicsResponse = childControllerClient.getDeletableStoreTopics();
-        Assert.assertFalse(storeTopicsResponse.isError());
-        Assert.assertEquals(storeTopicsResponse.getTopics().size(), 0);
-      });
-    }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT, priority = 3)
-  public void testBootstrappingVeniceChangelogConsumer_MultiConsumersWithPollAndCatchUp() throws Exception {
-    ControllerClient childControllerClient =
-        new ControllerClient(clusterName, childDatacenters.get(0).getControllerConnectString());
-    // create a active-active enabled store and run batch push job
-    // batch job contains 100 records
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = TestWriteUtils.writeSimpleAvroFileWithStringToStringSchema(inputDir);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String storeName = Utils.getUniqueString("store");
-    Properties props =
-        TestWriteUtils.defaultVPJProps(parentControllers.get(0).getControllerUrl(), inputDirPath, storeName);
-    String keySchemaStr = recordSchema.getField(DEFAULT_KEY_FIELD_PROP).schema().toString();
-    String valueSchemaStr = recordSchema.getField(DEFAULT_VALUE_FIELD_PROP).schema().toString();
-    Map<String, String> viewConfig = new HashMap<>();
-    viewConfig.put(
-        "testView",
-        "{\"viewClassName\" : \"" + TestView.class.getCanonicalName() + "\", \"viewParameters\" : {}}");
-
-    viewConfig.put(
-        "changeCaptureView",
-        "{\"viewClassName\" : \"" + ChangeCaptureView.class.getCanonicalName() + "\", \"viewParameters\" : {}}");
-    UpdateStoreQueryParams storeParms = new UpdateStoreQueryParams().setActiveActiveReplicationEnabled(true)
-        .setHybridRewindSeconds(500)
-        .setHybridOffsetLagThreshold(8)
-        .setChunkingEnabled(true)
-        .setNativeReplicationEnabled(true)
-        .setPartitionCount(3);
-    MetricsRepository metricsRepository = new MetricsRepository();
-    ControllerClient setupControllerClient =
-        createStoreForJob(clusterName, keySchemaStr, valueSchemaStr, props, storeParms);
-    storeParms.setStoreViews(viewConfig);
-    // IntegrationTestPushUtils.updateStore(clusterName, props, storeParms);
-    setupControllerClient
-        .retryableRequest(5, controllerClient1 -> setupControllerClient.updateStore(storeName, storeParms));
-    // controllerClient.updateStore(storeName, storeParms);
-    TestWriteUtils.runPushJob("Run push job", props);
-    Map<String, String> samzaConfig = getSamzaConfig(storeName);
-    VeniceSystemFactory factory = new VeniceSystemFactory();
-    // Use a unique key for DELETE with RMD validation
-    int deleteWithRmdKeyIndex = 1000;
-
-    TestMockTime testMockTime = new TestMockTime();
-    ZkServerWrapper localZkServer = multiRegionMultiClusterWrapper.getChildRegions().get(0).getZkServerWrapper();
-    try (PubSubBrokerWrapper localKafka = ServiceFactory.getPubSubBroker(
-        new PubSubBrokerConfigs.Builder().setZkWrapper(localZkServer)
-            .setMockTime(testMockTime)
-            .setRegionName("local-pubsub")
-            .build())) {
-      Properties consumerProperties = new Properties();
-      String localKafkaUrl = localKafka.getAddress();
-      consumerProperties.put(KAFKA_BOOTSTRAP_SERVERS, localKafkaUrl);
-      consumerProperties.put(CLUSTER_NAME, clusterName);
-      consumerProperties.put(ZOOKEEPER_ADDRESS, localZkServer.getAddress());
-      ChangelogClientConfig globalChangelogClientConfig = new ChangelogClientConfig().setViewName("changeCaptureView")
-          .setConsumerProperties(consumerProperties)
-          .setControllerD2ServiceName(D2_SERVICE_NAME)
-          .setD2ServiceName(VeniceRouterWrapper.CLUSTER_DISCOVERY_D2_SERVICE_NAME)
-          .setLocalD2ZkHosts(localZkServer.getAddress())
-          .setControllerRequestRetryCount(3)
-          .setBootstrapFileSystemPath(Utils.getUniqueString(inputDirPath))
-          .setDatabaseSyncBytesInterval(100);
-      VeniceChangelogConsumerClientFactory veniceChangelogConsumerClientFactory =
-          new VeniceChangelogConsumerClientFactory(globalChangelogClientConfig, metricsRepository);
-
-      BootstrappingVeniceChangelogConsumer<Utf8, Utf8> bootstrappingVeniceChangelogConsumer_1 =
-          veniceChangelogConsumerClientFactory.getBootstrappingChangelogConsumer(storeName, "1");
-      BootstrappingVeniceChangelogConsumer<Utf8, Utf8> bootstrappingVeniceChangelogConsumer_2 =
-          veniceChangelogConsumerClientFactory.getBootstrappingChangelogConsumer(storeName, "2");
-      BootstrappingVeniceChangelogConsumer<Utf8, Utf8> bootstrappingVeniceChangelogConsumer_3 =
-          veniceChangelogConsumerClientFactory.getBootstrappingChangelogConsumer(storeName, "3");
-      try (VeniceSystemProducer veniceProducer =
-          factory.getClosableProducer("venice", new MapConfig(samzaConfig), null)) {
-        veniceProducer.start();
-        // Run Samza job to send PUT and DELETE requests.
-        runSamzaStreamJob(veniceProducer, storeName, null, 10, 10, 100);
-        // Produce a DELETE record with large timestamp
-        produceRecordWithLogicalTimestamp(veniceProducer, storeName, deleteWithRmdKeyIndex, 1000, true);
-      }
-
-      try (AvroGenericStoreClient<String, Utf8> client = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName)
-              .setVeniceURL(clusterWrapper.getRandomRouterURL())
-              .setMetricsRepository(metricsRepository))) {
-        TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-          Assert.assertNull(client.get(Integer.toString(deleteWithRmdKeyIndex)).get());
-        });
-      }
-
-      // Wait for 10 seconds so bootstrap can load results from kafka
-      Utils.sleep(10000);
-      bootstrappingVeniceChangelogConsumer_1.start(ImmutableSet.of(0)).get();
-      bootstrappingVeniceChangelogConsumer_2.start(ImmutableSet.of(1)).get();
-      bootstrappingVeniceChangelogConsumer_3.start(ImmutableSet.of(2)).get();
-      Map<String, PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> polledChangeEvents = new HashMap<>();
-
-      // 21 changes in nearline. 10 puts, 10 deletes and 1 delete with timestamp will be ignored and 1 end of bootstrap
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, bootstrappingVeniceChangelogConsumer_1);
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, bootstrappingVeniceChangelogConsumer_2);
-        pollChangeEventsFromChangeCaptureConsumer(polledChangeEvents, bootstrappingVeniceChangelogConsumer_3);
-        Assert.assertEquals(polledChangeEvents.size(), 11);
-
-        for (int i = 100; i < 110; i++) {
-          String key = Integer.toString(i);
-          ChangeEvent<Utf8> changeEvent = polledChangeEvents.get(key).getValue();
-          Assert.assertNotNull(changeEvent);
-          Assert.assertNull(changeEvent.getPreviousValue());
-          Assert.assertEquals(changeEvent.getCurrentValue().toString(), "stream_" + i);
-        }
-
-        Assert.assertTrue(polledChangeEvents.containsKey(null));
-        Assert.assertTrue(polledChangeEvents.get(null).isEndOfBootstrap());
-      });
-
-      polledChangeEvents.clear();
-
-      // Since nothing is produced, so no changed events generated.
-      TestUtils.waitForNonDeterministicAssertion(5, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumer2(polledChangeEvents, bootstrappingVeniceChangelogConsumer_1);
-        pollChangeEventsFromChangeCaptureConsumer2(polledChangeEvents, bootstrappingVeniceChangelogConsumer_2);
-        pollChangeEventsFromChangeCaptureConsumer2(polledChangeEvents, bootstrappingVeniceChangelogConsumer_3);
-        Assert.assertEquals(polledChangeEvents.size(), 0);
-      });
-
-      // Verify total updates match up
-      TestUtils.waitForNonDeterministicAssertion(
-          5,
-          TimeUnit.SECONDS,
-          () -> Assert.assertEquals(TestView.getInstance().getRecordCountForStore(storeName), 21));
-      parentControllerClient.disableAndDeleteStore(storeName);
-      // Verify that topics and store is cleaned up
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-        MultiStoreTopicsResponse storeTopicsResponse = childControllerClient.getDeletableStoreTopics();
-        Assert.assertFalse(storeTopicsResponse.isError());
-        Assert.assertEquals(storeTopicsResponse.getTopics().size(), 0);
-      });
-    }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT)
-  public void testAfterImageChangelogConsumer() throws Exception {
-    int batchRecordCount = 100;
-    int hybridRecordCount = 10;
-    int hybridRecordStartIndex = 100;
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = TestWriteUtils.writeSimpleAvroFileWithStringToStringSchema(inputDir);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String storeName = Utils.getUniqueString("store");
-    Properties props =
-        TestWriteUtils.defaultVPJProps(parentControllers.get(0).getControllerUrl(), inputDirPath, storeName);
-    String keySchemaStr = recordSchema.getField(DEFAULT_KEY_FIELD_PROP).schema().toString();
-    String valueSchemaStr = recordSchema.getField(DEFAULT_VALUE_FIELD_PROP).schema().toString();
-    UpdateStoreQueryParams storeParms = new UpdateStoreQueryParams().setActiveActiveReplicationEnabled(true)
-        .setHybridRewindSeconds(500)
-        .setHybridOffsetLagThreshold(8)
-        .setChunkingEnabled(true)
-        .setNativeReplicationEnabled(true)
-        .setPartitionCount(1);
-    MetricsRepository metricsRepository = new MetricsRepository();
-    ControllerClient setupControllerClient =
-        createStoreForJob(clusterName, keySchemaStr, valueSchemaStr, props, storeParms);
-    setupControllerClient
-        .retryableRequest(5, controllerClient1 -> setupControllerClient.updateStore(storeName, storeParms));
-    TestWriteUtils.runPushJob("Run push job", props);
-    Map<String, String> samzaConfig = getSamzaConfig(storeName);
-    VeniceSystemFactory factory = new VeniceSystemFactory();
-
-    TestMockTime testMockTime = new TestMockTime();
-    ZkServerWrapper localZkServer = multiRegionMultiClusterWrapper.getChildRegions().get(0).getZkServerWrapper();
-    try (PubSubBrokerWrapper localKafka = ServiceFactory.getPubSubBroker(
-        new PubSubBrokerConfigs.Builder().setZkWrapper(localZkServer)
-            .setMockTime(testMockTime)
-            .setRegionName("local-pubsub")
-            .build())) {
-      Properties consumerProperties = new Properties();
-      String localKafkaUrl = localKafka.getAddress();
-      consumerProperties.put(KAFKA_BOOTSTRAP_SERVERS, localKafkaUrl);
-      consumerProperties.put(CLUSTER_NAME, clusterName);
-      consumerProperties.put(ZOOKEEPER_ADDRESS, localZkServer.getAddress());
-      ChangelogClientConfig globalChangelogClientConfig =
-          new ChangelogClientConfig().setConsumerProperties(consumerProperties)
-              .setControllerD2ServiceName(D2_SERVICE_NAME)
-              .setD2ServiceName(VeniceRouterWrapper.CLUSTER_DISCOVERY_D2_SERVICE_NAME)
-              .setLocalD2ZkHosts(localZkServer.getAddress())
-              .setControllerRequestRetryCount(3)
-              .setBootstrapFileSystemPath(Utils.getUniqueString(inputDirPath));
-      VeniceChangelogConsumerClientFactory veniceChangelogConsumerClientFactory =
-          new VeniceChangelogConsumerClientFactory(globalChangelogClientConfig, metricsRepository);
-
-      VeniceChangelogConsumer<Utf8, Utf8> changelogConsumer =
-          veniceChangelogConsumerClientFactory.getChangelogConsumer(storeName);
-      try (VeniceSystemProducer veniceProducer =
-          factory.getClosableProducer("venice", new MapConfig(samzaConfig), null)) {
-        veniceProducer.start();
-        // Run Samza job to send PUT
-        runSamzaStreamJob(veniceProducer, storeName, null, hybridRecordCount, 0, hybridRecordStartIndex);
-      }
-
-      try (AvroGenericStoreClient<String, Utf8> client = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName)
-              .setVeniceURL(clusterWrapper.getRandomRouterURL())
-              .setMetricsRepository(metricsRepository))) {
-        TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-          Assert.assertNotNull(client.get(Integer.toString(hybridRecordStartIndex + hybridRecordCount - 1)).get());
-        });
-      }
-      changelogConsumer.subscribe(Collections.singleton(0)).get();
-
-      List<PubSubMessage<Utf8, ChangeEvent<Utf8>, VeniceChangeCoordinate>> changedEventList = new ArrayList<>();
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-        pollChangeEventsFromChangeCaptureConsumerToList(changedEventList, changelogConsumer);
-        Assert.assertEquals(changedEventList.size(), batchRecordCount + hybridRecordCount);
-      });
+  private void produceRecordWithLogicalTimestamp(
+      VeniceSystemProducer veniceProducer,
+      String storeName,
+      int index,
+      long logicalTimestamp,
+      boolean isDeleteOperation) {
+    if (isDeleteOperation) {
+      sendStreamingDeleteRecord(veniceProducer, storeName, Integer.toString(index), logicalTimestamp);
+    } else {
+      sendStreamingRecord(veniceProducer, storeName, Integer.toString(index), "stream_" + index, logicalTimestamp);
     }
   }
 
@@ -1581,19 +514,6 @@ public class TestActiveActiveIngestion {
           storeName,
           Integer.toString(i),
           mockedTime == null ? null : mockedTime.getMilliseconds());
-    }
-  }
-
-  private void produceRecordWithLogicalTimestamp(
-      VeniceSystemProducer veniceProducer,
-      String storeName,
-      int index,
-      long logicalTimestamp,
-      boolean isDeleteOperation) {
-    if (isDeleteOperation) {
-      sendStreamingDeleteRecord(veniceProducer, storeName, Integer.toString(index), logicalTimestamp);
-    } else {
-      sendStreamingRecord(veniceProducer, storeName, Integer.toString(index), "stream_" + index, logicalTimestamp);
     }
   }
